API Move convertVariableNames to utils.py
