+++ serpentTools/utils.py
@@ -97,3 +97,14 @@ def splitValsUncs(iterable, copy=False):
     if copy:
         return vals.copy(), uncs.copy()
     return vals, uncs
+
+
+def convertVariableName(variable):
+    """Convert a SERPENT variable to camelCase"""
+    lowerSplits = [item.lower() for item in variable.split('_')]
+    if len(lowerSplits) == 1:
+        return lowerSplits[0]
+    else:
+        return lowerSplits[0] + ''.join([item.capitalize()
+                                         for item in lowerSplits[1:]])
+
